refactor(decision): move ActionProposal out of maiw-mcp
ActionProposal and RiskLevel had no dependency on the MCP SDK — only on
uuid, datetime, and pydantic. They belong with the decision layer, not
the transport layer.

Changes:
- Add packages/maiw-decision/maiw_decision/proposal.py (canonical home)
- Export ActionProposal + RiskLevel from maiw_decision.__init__
- Update maiw_decision/models.py + engine.py to import from new path
- Update all consumers (maiw-execution, maiw-skills, maiw-agents,
  apps/api, mcp_servers, src/api) to import from maiw_decision.proposal
- Drop maiw-mcp dependency from maiw-decision/pyproject.toml
- Empty out maiw_mcp/contracts/actions.py (tombstone comment only)

Runtime behavior: unchanged. Test baseline: 9 failed (pre-existing), 167 passed.
