+++ packages/maiw-decision/pyproject.toml
@@ -9,7 +9,6 @@ description = "Deterministic decision engine for MAIW warehouse write operations
 requires-python = ">=3.11"
 dependencies = [
     "pydantic>=2.0.0",
-    "maiw-mcp>=0.1.0",
     "maiw-state>=0.1.0",
 ]
