+++ packages/maiw-execution/maiw_execution/base.py
@@ -57,7 +57,7 @@ from typing import Any, Optional, Protocol, runtime_checkable
 from pydantic import BaseModel, Field, model_validator
 
 from maiw_decision.models import DecisionOutcome, DecisionResult
-from maiw_mcp.contracts.actions import ActionProposal
+from maiw_decision.proposal import ActionProposal
 from maiw_mcp.deadline import RequestDeadline, RequestDeadlineExceeded
 
 from .outcome import AmbiguousWriteError, ExecutionOutcome
