+++ packages/maiw-skills/maiw_skills/equipment/skills.py
@@ -29,7 +29,7 @@ import logging
 import os
 
 from maiw_mcp.client.client import MAIWMCPClient
-from maiw_mcp.contracts.actions import ActionProposal
+from maiw_decision.proposal import ActionProposal
 from maiw_mcp.contracts.equipment import (
     EQUIPMENT_ASSIGN_METADATA,
     EQUIPMENT_GET_STATUS_METADATA,
