+++ packages/maiw-skills/maiw_skills/labor/skills.py
@@ -25,7 +25,7 @@ import logging
 import os
 
 from maiw_mcp.client.client import MAIWMCPClient
-from maiw_mcp.contracts.actions import ActionProposal
+from maiw_decision.proposal import ActionProposal
 from maiw_mcp.contracts.labor import (
     LABOR_ALLOCATE_METADATA,
     LABOR_GET_ALLOCATION_METADATA,
