+++ src/api/agents/operations/labor_executor.py
@@ -26,7 +26,7 @@ from datetime import datetime, timezone
 from typing import Any, Optional
 
 from maiw_decision.models import DecisionOutcome, DecisionResult
-from maiw_mcp.contracts.actions import ActionProposal
+from maiw_decision.proposal import ActionProposal
 from src.api.agents.inventory.action_executor import (
     ActionConflict,
     ActionDecisionMismatch,
