+++ tests/contract/test_equipment_capability.py
@@ -19,7 +19,7 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from maiw_mcp.contracts.actions import ActionProposal, RiskLevel
+from maiw_decision.proposal import ActionProposal, RiskLevel
 from maiw_mcp.contracts.equipment import (
     EQUIPMENT_ASSIGN_METADATA,
     EQUIPMENT_GET_STATUS_METADATA,
