+++ tests/contract/test_wave_capability.py
@@ -17,7 +17,7 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 
-from maiw_mcp.contracts.actions import ActionProposal, RiskLevel
+from maiw_decision.proposal import ActionProposal, RiskLevel
 from maiw_mcp.contracts.wave import (
     WAVE_GET_METADATA,
     WAVE_GET_RISK_METADATA,
