+++ tests/mcp/test_decision_integration.py
@@ -34,7 +34,7 @@ from maiw_decision import (
     DecisionOutcome,
     DecisionRequest,
 )
-from maiw_mcp.contracts.actions import ActionProposal, RiskLevel
+from maiw_decision.proposal import ActionProposal, RiskLevel
 from maiw_mcp.contracts.equipment import (
     EquipmentAssetInfo,
     EquipmentStatusRequest,
