+++ tests/unit/reliability/test_idempotency.py
@@ -27,7 +27,7 @@ from maiw_execution import (
     ExecutionRegistry,
     LaborActionExecutor,
 )
-from maiw_mcp.contracts.actions import ActionProposal, RiskLevel
+from maiw_decision.proposal import ActionProposal, RiskLevel
 from maiw_mcp.contracts.labor import LaborAllocateResult
