+++ tests/unit/test_decision_engine.py
@@ -34,7 +34,7 @@ from maiw_decision import (
     DecisionRequest,
     DecisionResult,
 )
-from maiw_mcp.contracts.actions import ActionProposal, RiskLevel
+from maiw_decision.proposal import ActionProposal, RiskLevel
 from maiw_state import (
     EquipmentAssetSummary,
     EquipmentState,
