+++ tests/unit/test_package_imports.py
@@ -98,7 +98,7 @@ class TestPackageImportSmoke:
         assert WaveState is not None
 
     def test_canonical_mcp_import(self):
-        from maiw_mcp.contracts.actions import ActionProposal
+        from maiw_decision.proposal import ActionProposal
         from maiw_mcp.contracts.inventory import INVENTORY_GET_METADATA
         from maiw_mcp.contracts.equipment import EQUIPMENT_ASSIGN_METADATA
         from maiw_mcp.contracts.labor import LABOR_ALLOCATE_METADATA
