+++ tests/unit/test_state_aware_ops_phase6.py
@@ -25,7 +25,7 @@ from unittest.mock import AsyncMock, MagicMock, call
 import pytest
 
 from maiw_decision import DecisionEngine
-from maiw_mcp.contracts.actions import ActionProposal, RiskLevel
+from maiw_decision.proposal import ActionProposal, RiskLevel
 from maiw_mcp.contracts.equipment import (
     EquipmentAssetInfo,
     EquipmentExecuteAssignResult,
